requirements.txt
streamlit
requests
deep_translator
Pillow